Work in progress: Migrate Dockerfiles from clamav + clambc
We have found a need to be able to easily update the Dockerfile and
scripts without publishing new versions of the ClamAV or compiler
projects.

The layout introduced in this commit is like this:

project/version/platform

Where "unstable" tracks the main (development) branch.

A Jenkinsfile and Dockerfile are placed in each "platform" directory.