Build dependencies from source
Build the library dependencies from source using Mussels.
The reason is that `alpine:latest` is still on Rust version 1.60.
Several Rust dependencies now require Rust version 1.61+ including three
that fix panics when opening images. So it is not a viable option to pin
the dependencies to older version to lower the Rust version requirement.

Instead, we're building with Rust from rustup. HOWEVER, changes in 1.59
now build all Musl targetes with the -static-pie option which means it
cannot link with the `.so` libraries. Their intention is to prevent
accidentally linking with native shared libs of the wrong architecture
when cross compiling. But that broke things for us because Alpine does
not provide static libs for all of our dependencies. Fortunately we have
Mussels recipes to build those, so it's relatively simple for us to do
that ourselves.

The only exception is libmilter. We do not (yet) have a recipe to build
libmilter. So at present the images built with this do not include
clamav-milter.