Revert change to use /tmp instead of /var/lock or /var/clamav
This change breaks things for existing ClamAV users.

We will need to find another solution such as altering these config
options in a downstream Dockerfile for building the unprivileged image.

Also bring back libmilter and clamav-milter.
Alpine provides a static lib (libmilter.a) which means we don't *have*
to build one with Mussels so we don't have to wait until a recipe can be
made for libmilter.

Note: In this 'revert' we're still changing ownership of the database
directory in the Dockerfile as well as now in the (privileged)
entrypoint. The reason is that if it is a mounted directory, changing
ownership in the Dockerfile may have had no effect and it has to be done
on start-up. Additional research needed.