Fix compatibility with custom `.conf`'s that use `/run/...` instead of `/tmp/...`
Provide `/run/lock` and `/run/clamav` directories for users that are
providing their own config files and expect the those paths to work.
And on startup, look for `clamd.sock` in either `/var/clamav` or `/tmp`,
for compatibility.

Note: In this commit we're now changing ownership of the database
directory in the Dockerfile as well as now in the (privileged)
entrypoint script. 
The reason is that if it is a mounted directory, changing ownership in
the Dockerfile may have had no effect and it has to be done on start-up.
Additional research required to support use case when adding the
unprivileged image.

This commit also brings back libmilter and clamav-milter.
Alpine provides a static lib (libmilter.a) which means we don't *have*
to build one with Mussels so we don't have to wait until a recipe can be
made for libmilter.

Also bump the 0.105.1 image revision from 5 -> 6.