Add ability to set application config using env vars
As per existing behaviour we are not able to change the ClamD and
Freshclam when we run clamav docker containers.

To fix this, user can now provide env variables at runtime which
will be used to update corresponding variables in ClamD and freshclam
config files. For naming conventions of env variables & details,
check the README files.

Credit for these changes goes to https://github.com/Loki-Afro

Fixes: https://github.com/Cisco-Talos/clamav-docker/issues/27