Remove PidFile configuration
There should be no need for freshclam, clamd, or clamav-milter to create
a PID file.

There is a very minor security concern that the PID file is located in a
user-writable directory in that a naive service manager might kill the
wrong process if the PID file has been modified.

Granted, these Docker images do not have a service manager which
operates in this way. However, I believe there is no harm in removing
the PID files so as to eliminate the hypothetical risk if one were to
use those PID files.

Refs:
- https://github.com/Cisco-Talos/clamav/issues/1076
- https://github.com/Cisco-Talos/clamav/commit/5b168b50cdd0c0650eca3047097a3f6919e8ab64