Migrate multi-arch build away from macOS for stability
MacOS arm64 to amd64 builds are unstable and began to fail when
calcualting SHA hashes on files greater than 4095-bytes in length.
This is not a ClamAV bug, or even an OpenSSL bug as far as I can tell.
The build and tests work fine using Docker buildx from Ubuntu.

CLAM-2855